// src/types/anilist.ts
export type Maybe<T> = T | null;
// ...
export interface MediaConnection {
  edges: Maybe<MediaEdge[]>;
  nodes: Maybe<Media[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}

/** Media connection edge */
export interface MediaEdge {
  node: Maybe<Media>;
  /** The id of the connection*/
  id: Maybe<number>;
  /** The type of relation to the parent model*/
  relationType: Maybe<MediaRelation>;
  /** If the studio is the main animation studio of the media (For Studio->MediaConnection field only)*/
  isMainStudio: boolean;
  /** The characters in the media voiced by the parent actor*/
  characters: Maybe<Character[]>;
  /** The characters role in the media*/
  characterRole: Maybe<CharacterRole>;
  /** Media specific character name*/
  characterName: Maybe<string>;
  /** Notes regarding the VA's role for the character*/
  roleNotes: Maybe<string>;
  /** Used for grouping roles where multiple dubs exist for the same language. Either dubbing company name or language variant.*/
  dubGroup: Maybe<string>;
  /** The role of the staff member in the production of the media*/
  staffRole: Maybe<string>;
  /** The voice actors of the character*/
  voiceActors: Maybe<Staff[]>;
  /** The voice actors of the character with role date*/
  voiceActorRoles: Maybe<StaffRoleType[]>;
  /** The order the media should be displayed from the users favourites*/
  favouriteOrder: Maybe<number>;
}

/** Anime or Manga */
export interface Media {
  /** The id of the media*/
  id: number;
  /** The mal id of the media*/
  idMal: Maybe<number>;
  /** The official titles of the media in various languages*/
  title: Maybe<MediaTitle>;
  /** The type of the media; anime or manga*/
  type: Maybe<MediaType>;
  /** The format the media was released in*/
  format: Maybe<MediaFormat>;
  /** The current releasing status of the media*/
  status: Maybe<MediaStatus>;
  /** Short description of the media's story and characters*/
  description: Maybe<string>;
  /** The first official release date of the media*/
  startDate: Maybe<FuzzyDate>;
  /** The last official release date of the media*/
  endDate: Maybe<FuzzyDate>;
  /** The season the media was initially released in*/
  season: Maybe<MediaSeason>;
  /** The season year the media was initially released in*/
  seasonYear: Maybe<number>;
  /** The year & season the media was initially released in*/
  seasonInt: Maybe<number>;
  /** The amount of episodes the anime has when complete*/
  episodes: Maybe<number>;
  /** The general length of each anime episode in minutes*/
  duration: Maybe<number>;
  /** The amount of chapters the manga has when complete*/
  chapters: Maybe<number>;
  /** The amount of volumes the manga has when complete*/
  volumes: Maybe<number>;
  /** Where the media was created. (ISO 3166-1 alpha-2)*/
  countryOfOrigin: Maybe<string>;
  /** If the media is officially licensed or a self-published doujin release*/
  isLicensed: Maybe<boolean>;
  /** Source type the media was adapted from.*/
  source: Maybe<MediaSource>;
  /** Official Twitter hashtags for the media*/
  hashtag: Maybe<string>;
  /** Media trailer or advertisement*/
  trailer: Maybe<MediaTrailer>;
  /** When the media's data was last updated*/
  updatedAt: Maybe<number>;
  /** The cover images of the media*/
  coverImage: Maybe<MediaCoverImage>;
  /** The banner image of the media*/
  bannerImage: Maybe<string>;
  /** The genres of the media*/
  genres: Maybe<string[]>;
  /** Alternative titles of the media*/
  synonyms: Maybe<string[]>;
  /** A weighted average score of all the user's scores of the media*/
  averageScore: Maybe<number>;
  /** Mean score of all the user's scores of the media*/
  meanScore: Maybe<number>;
  /** The number of users with the media on their list*/
  popularity: Maybe<number>;
  /** Locked media may not be added to lists our favorited. This may be due to the entry pending for deletion or other reasons.*/
  isLocked: Maybe<boolean>;
  /** The amount of related activity in the past hour*/
  trending: Maybe<number>;
  /** The amount of user's who have favourited the media*/
  favourites: Maybe<number>;
  /** List of tags that describes elements and themes of the media*/
  tags: Maybe<MediaTag[]>;
  /** Other media in the same or connecting franchise*/
  relations: Maybe<MediaConnection>;
  /** The characters in the media*/
  characters: Maybe<CharacterConnection>;
  /** The staff who produced the media*/
  staff: Maybe<StaffConnection>;
  /** The companies who produced the media*/
  studios: Maybe<StudioConnection>;
  /** If the media is marked as favourite by the current authenticated user*/
  isFavourite: boolean;
  /** If the media is blocked from being added to favourites*/
  isFavouriteBlocked: boolean;
  /** If the media is intended only for 18+ adult audiences*/
  isAdult: Maybe<boolean>;
  /** The media's next episode airing schedule*/
  nextAiringEpisode: Maybe<AiringSchedule>;
  /** The media's entire airing schedule*/
  airingSchedule: Maybe<AiringScheduleConnection>;
  /** The media's daily trend stats*/
  trends: Maybe<MediaTrendConnection>;
  /** External links to another site related to the media*/
  externalLinks: Maybe<MediaExternalLink[]>;
  /** Data and links to legal streaming episodes on external sites*/
  streamingEpisodes: Maybe<MediaStreamingEpisode[]>;
  /** The ranking of the media in a particular time span and format compared to other media*/
  rankings: Maybe<MediaRank[]>;
  /** The authenticated user's media list entry for the media*/
  mediaListEntry: Maybe<MediaList>;
  /** User recommendations for similar media*/
  recommendations: Maybe<RecommendationConnection>;
  /** The url for the media page on the AniList website*/
  siteUrl: Maybe<string>;
  /** If the media should have forum thread automatically created for it on airing episode release*/
  autoCreateForumThread: Maybe<boolean>;
  /** If the media is blocked from being recommended to/from*/
  isRecommendationBlocked: Maybe<boolean>;
  /** If the media is blocked from being reviewed*/
  isReviewBlocked: Maybe<boolean>;
  /** Notes for site moderators*/
  modNotes: Maybe<string>;

  translations?: Translation[];

  action: any;
  name: any;
  email: any;
  role: any;
  createdAt: any;
}
// ...
export interface MediaTitle extends Record<string, string> {
  /** The romanization of the native language title*/
  romaji: Maybe<string>;
  /** The official english title*/
  english: Maybe<string>;
  /** Official title in it's native language*/
  native: Maybe<string>;
  /** The currently authenticated users preferred title language. Default romaji for non-authenticated*/
  userPreferred: Maybe<string>;
}

/** Media type enum, anime or manga. */
export enum MediaType {
  Anime = "ANIME",
  Manga = "MANGA",
}
/** The format the media was released in */
export enum MediaFormat {
  Tv = "TV",
  Tv_short = "TV_SHORT",
  Movie = "MOVIE",
  Special = "SPECIAL",
  Ova = "OVA",
  Ona = "ONA",
  Music = "MUSIC",
  Manga = "MANGA",
  Novel = "NOVEL",
  One_shot = "ONE_SHOT",
}
/** The format the language was released in */
export enum LanguagesDetectFormat {
  jpn = "ja",
  eng = "en",
  vie = "vi",
  chi_sim = "zh",
}
/** The format the language was released in */
export enum RevertLanguagesDetectFormat {
  ja = "jpn",
  en = "eng",
  vi = "vie",
  zh = "chi_sim",
}
/** The current releasing status of the media */
export enum MediaStatus {
  Finished = "FINISHED",
  Releasing = "RELEASING",
  Not_yet_released = "NOT_YET_RELEASED",
  Cancelled = "CANCELLED",
  Hiatus = "HIATUS",
}
/** Date object that allows for incomplete date values (fuzzy) */
export interface FuzzyDate {
  /** Numeric Year (2017)*/
  year: Maybe<number>;
  /** Numeric Month (3)*/
  month: Maybe<number>;
  /** Numeric Day (24)*/
  day: Maybe<number>;
}

export enum MediaSeason {
  Winter = "WINTER",
  Spring = "SPRING",
  Summer = "SUMMER",
  Fall = "FALL",
}

/** Source type the media was adapted from */
export enum MediaSource {
  Original = "ORIGINAL",
  Manga = "MANGA",
  Light_novel = "LIGHT_NOVEL",
  Visual_novel = "VISUAL_NOVEL",
  Video_game = "VIDEO_GAME",
  Other = "OTHER",
  Novel = "NOVEL",
  Doujinshi = "DOUJINSHI",
  Anime = "ANIME",
  Web_novel = "WEB_NOVEL",
  Live_action = "LIVE_ACTION",
  Game = "GAME",
  Comic = "COMIC",
  Multimedia_project = "MULTIMEDIA_PROJECT",
  Picture_book = "PICTURE_BOOK",
}
/** Media trailer or advertisement */
export interface MediaTrailer {
  /** The trailer video id*/
  id: Maybe<string>;
  /** The site the video is hosted by (Currently either youtube or dailymotion)*/
  site: Maybe<string>;
  /** The url for the thumbnail image of the video*/
  thumbnail: Maybe<string>;
}

export interface MediaCoverImage {
  /** The cover image url of the media at its largest size. If this size isn't available, large will be provided instead.*/
  extraLarge: Maybe<string>;
  /** The cover image url of the media at a large size*/
  large: Maybe<string>;
  /** The cover image url of the media at medium size*/
  medium: Maybe<string>;
  /** Average #hex color of cover image*/
  color: Maybe<string>;
}

/** A tag that describes a theme or element of the media */
export interface MediaTag {
  /** The id of the tag*/
  id: number;
  /** The name of the tag*/
  name: string;
  /** A general description of the tag*/
  description: Maybe<string>;
  /** The categories of tags this tag belongs to*/
  category: Maybe<string>;
  /** The relevance ranking of the tag out of the 100 for this media*/
  rank: Maybe<number>;
  /** If the tag could be a spoiler for any media*/
  isGeneralSpoiler: Maybe<boolean>;
  /** If the tag is a spoiler for this media*/
  isMediaSpoiler: Maybe<boolean>;
  /** If the tag is only for adult 18+ media*/
  isAdult: Maybe<boolean>;
  /** The user who submitted the tag*/
  userId: Maybe<number>;
}
// ...
export interface CharacterConnection {
  edges: Maybe<CharacterEdge[]>;
  nodes: Maybe<Character[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}
// ...
export interface StaffConnection {
  edges: Maybe<StaffEdge[]>;
  nodes: Maybe<Staff[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}
// ...
export interface StudioConnection {
  edges: Maybe<StudioEdge[]>;
  nodes: Maybe<Studio[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}
// ...
export interface AiringSchedule {
  /** The id of the airing schedule item*/
  id: number;
  /** The time the episode airs at*/
  airingAt: number;
  /** Seconds until episode starts airing*/
  timeUntilAiring: number;
  /** The airing episode number*/
  episode: number;
  /** The associate media id of the airing episode*/
  mediaId: number;
  /** The associate media of the airing episode*/
  media: Maybe<Media>;
}

export interface AiringScheduleConnection {
  edges: Maybe<AiringScheduleEdge[]>;
  nodes: Maybe<AiringSchedule[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}
// ...
export interface MediaTrendConnection {
  edges: Maybe<MediaTrendEdge[]>;
  nodes: Maybe<MediaTrend[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}
// ...
export interface MediaExternalLink {
  /** The id of the external link*/
  id: number;
  /** The url of the external link or base url of link source*/
  url: Maybe<string>;
  /** The links website site name*/
  site: string;
  /** The links website site id*/
  siteId: Maybe<number>;
  type: Maybe<ExternalLinkType>;
  /** Language the site content is in. See Staff language field for values.*/
  language: Maybe<string>;
  color: Maybe<string>;
  /** The icon image url of the site. Not available for all links. Transparent PNG 64x64*/
  icon: Maybe<string>;
  notes: Maybe<string>;
  isDisabled: Maybe<boolean>;
}

export enum ExternalLinkType {
  Info = "INFO",
  Streaming = "STREAMING",
  Social = "SOCIAL",
}
/** Data and links to legal streaming episodes on external sites */
export interface MediaStreamingEpisode {
  /** Title of the episode*/
  title: Maybe<string>;
  /** Url of episode image thumbnail*/
  thumbnail: Maybe<string>;
  /** The url of the episode*/
  url: Maybe<string>;
  /** The site location of the streaming episodes*/
  site: Maybe<string>;
}

/** The ranking of a media in a particular time span and format compared to other media */
export interface MediaRank {
  /** The id of the rank*/
  id: number;
  /** The numerical rank of the media*/
  rank: number;
  /** The type of ranking*/
  type: MediaRankType;
  /** The format the media is ranked within*/
  format: MediaFormat;
  /** The year the media is ranked within*/
  year: Maybe<number>;
  /** The season the media is ranked within*/
  season: Maybe<MediaSeason>;
  /** If the ranking is based on all time instead of a season/year*/
  allTime: Maybe<boolean>;
  /** String that gives context to the ranking type and time span*/
  context: string;
}

/** The type of ranking */
export enum MediaRankType {
  Rated = "RATED",
  Popular = "POPULAR",
}

/** Recommendation sort enums */
export enum RecommendationSort {
  Id = "ID",
  Id_desc = "ID_DESC",
  Rating = "RATING",
  Rating_desc = "RATING_DESC",
}
export interface RecommendationConnection {
  edges: Maybe<RecommendationEdge[]>;
  nodes: Maybe<Recommendation[]>;
  /** The pagination information*/
  pageInfo: Maybe<PageInfo>;
}